+++ src/components/body/Emoji.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { DataEmoji } from '../../dataUtils/DataTypes';
-import { emojiUrl } from '../../dataUtils/emojiSelectors';
+import { emojiUnified, emojiUrl } from '../../dataUtils/emojiSelectors';
 import './Emoji.css';
 
 type Props = Readonly<{
@@ -9,7 +9,7 @@ type Props = Readonly<{
 
 export function Emoji({ emoji }: Props) {
   return (
-    <button className="epr-emoji">
+    <button className="epr-emoji" data-unified={emojiUnified(emoji)}>
       <img src={emojiUrl(emoji)} className="epr-emoji-img" />
     </button>
   );
